WIP
Dashboard view now passes applicant list to DashboardList component

Cleaned up ESLint errors
